+++ src/constants/ExtensionState.ts
@@ -2,5 +2,6 @@
 export const ExtensionState = {
   PagesView: `frontMatter:Pages:ViewType`,
   SelectedFolder: `frontMatter:SelectedFolder`,
-  Version: `frontMatter:Version`
+  Version: `frontMatter:Version`,
+  SettingPromoted: `frontMatter:Settings:Promoted`
 };